move transforms logic to deprecated namespace
