+++ db/deprecated/transforms/operations/finish_specifying.py
@@ -2,7 +2,7 @@ import psycopg
 from sqlalchemy import inspect
 
 from db.tables.operations import select as tables_select
-from db.transforms.base import Summarize
+from db.deprecated.transforms.base import Summarize
 from db.deprecated.columns import get_column_from_oid_and_attnum
