+++ db/tests/deprecated/queries/test_base.py
@@ -1,7 +1,7 @@
 from sqlalchemy import inspect
 from db.deprecated.queries.base import DBQuery, InitialColumn, JoinParameter
 from db.deprecated.columns import get_column_attnum_from_name as get_attnum
-from db.transforms import base as tbase
+from db.deprecated.transforms import base as tbase
 from db.deprecated.metadata import get_empty_metadata
